1)Modified joystick and WS collecting codes, 2)Changed file names
