+++ .gitignore
@@ -1,1 +1,3 @@
 ee629wsproject-7d249e8f524e.json
+*.json
+!spec/*.json
